duplicates bug fixed

+++ files/upload/index.php
@@ -9,13 +9,18 @@ for( $i=0 ; $i < $total ; $i++ ) {
   $tmpFilePath = $_FILES['files']['tmp_name'][$i];
 
   if ($tmpFilePath != ""){
-    $newFilePath = "../uploads/" . $_FILES['files']['name'][$i];
+    $name = $_FILES['files']['name'][$i];
+	$newFilePath = "../uploads/" . $_FILES['files']['name'][$i];
 	$databasePath = "/files/uploads/".$_FILES['files']['name'][$i];
-
-    if(move_uploaded_file($tmpFilePath, $newFilePath)) {
-		$dbConnection->createFile($_FILES['files']['name'][$i], $databasePath);
-		echo "Plik został przesłany na serwer";
+	if(!file_exists($newFilePath)){
+		if(move_uploaded_file($tmpFilePath, $newFilePath)) {
+			$dbConnection->createFile($_FILES['files']['name'][$i], $databasePath);
+			echo "Plik $name został przesłany na serwer<br>";
 		}
-	}
+  } else {
+	
+	echo "Plik o nazwie $name istnieje na serwerze<br>";
+  }
+  }
 }
 ?>
